+++ src/main/java/com/appdynamics/extensions/jmx/metrics/processor/ListMetricsProcessor.java
@@ -1,10 +1,12 @@
-package com.appdynamics.extensions.jmx.metrics;
+package com.appdynamics.extensions.jmx.metrics.processor;
+
+import com.appdynamics.extensions.jmx.metrics.MetricDetails;
 
 import javax.management.Attribute;
 import java.util.List;
 
-import static com.appdynamics.extensions.jmx.JMXUtil.getMetricAfterCharacterReplacement;
-import static com.appdynamics.extensions.jmx.metrics.Constants.PERIOD;
+import static com.appdynamics.extensions.jmx.utils.JMXUtil.getMetricAfterCharacterReplacement;
+import static com.appdynamics.extensions.jmx.utils.Constants.PERIOD;
 
 /**
  * Created by bhuvnesh.kumar on 3/11/19.
